refactor: optimise package size
