+++ .gitignore
@@ -18,6 +18,7 @@
 
 # testing artifacts
 **/tests/playground/*.mjs
+**/*-test-artifact.*
 
 # moon
 .moon/cache
