+++ @kindspells/astro-shield/src/e2e/fixtures/dynamic/astro.config.mjs
