+++ @kindspells/astro-shield/src/e2e/fixtures/dynamic/package.json
