+++ @kindspells/astro-shield/src/e2e/fixtures/dynamic/src/env.d.ts
