+++ @kindspells/astro-shield/src/e2e/fixtures/dynamic/src/pages/index.astro
