+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid/astro.config.mjs
