+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid/package.json
