+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid/public/code.js
