+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid/src/env.d.ts
