+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid/src/pages/index.astro
