+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid2/public/code.js
