+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid2/src/env.d.ts
