+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid2/src/pages/index.astro
