+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid3/astro.config.mjs
