+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid3/public/code.js
