+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid3/src/env.d.ts
