+++ @kindspells/astro-shield/src/e2e/fixtures/hybrid3/src/pages/static.astro
