+++ @kindspells/astro-shield/src/e2e/fixtures/static/astro.config.mjs
