+++ @kindspells/astro-shield/src/e2e/fixtures/static/package.json
