+++ @kindspells/astro-shield/src/e2e/fixtures/static/src/env.d.ts
