+++ @kindspells/astro-shield/src/e2e/fixtures/static/src/pages/index.astro
