+++ @kindspells/astro-shield/vitest.config.e2e.mts
@@ -8,6 +8,6 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
 	test: {
-		include: ['e2e/**/*.test.mts'],
+		include: ['src/e2e/**/*.test.mts'],
 	},
 })
